Complete PostgreSQL migration and project cleanup.

+++ backend/app/models/office.py
@@ -1,6 +1,6 @@
 # app/models/office.py - An area of responsibility for a given body.
 
-from sqlalchemy import Column, Integer, String, Float, BigInteger, ForeignKey
+from sqlalchemy import Column, Integer, String, Float, ForeignKey
 from sqlalchemy.orm import relationship
 from app.database import Base
 
@@ -11,7 +11,8 @@ class Office(Base):
     office_id = Column(Integer, primary_key=True, autoincrement=True)
     title = Column(String(45), nullable=True, default=None)
     office_precedence = Column(Float, nullable=True, default=None)
-    office_body_id = Column(BigInteger, ForeignKey('body.body_id'), nullable=False)
+    # Use Integer to match Body.body_id (PostgreSQL requires FK type compatibility)
+    office_body_id = Column(Integer, ForeignKey('body.body_id'), nullable=False)
 
     # Relationships
     body = relationship("Body", back_populates="offices")
